Ruleの具象クラスをrules/以下に移動
